pyproject.toml
[project]
name = "pdf-to-text-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "openai==1.109.1",
    "openai-agents==0.3.3",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.7",
    "pinecone>=7.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.0",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.27",
    "supabase>=2.22.0",
    "tuspy>=1.1.0",
    "uvicorn>=0.37.0",
]
